fix: clippy auto-fix + manual cleanup of warnings
- Remove unknown lint attributes (clippy::sort_by_key) from 10 files
- map_or(false, ...) -> is_some_and(...) in rairos-mcp
- Elide unnecessary lifetimes in extract_tag
- Remove dead test helper temp_manager()
- Auto-fix from clippy --fix across 40+ files

+++ crates/rairos-review-generator/src/lib.rs
@@ -666,7 +666,7 @@ mod tests {
     fn test_build_timeline_truncation() {
         let gen = ReviewGenerator::new();
         let papers: Vec<PaperRef> = (0..25)
-            .map(|i| make_paper(&format!("p{}", i), &format!("Paper {}", i), "Abstract", 2020 + i as i32))
+            .map(|i| make_paper(&format!("p{}", i), &format!("Paper {}", i), "Abstract", 2020 + i))
             .collect();
         let timeline = gen.build_timeline(&papers);
         assert_eq!(timeline.len(), 20); // truncated to 20
